+++ src/components/app-wrapper/__spec__.js
@@ -1,15 +1,14 @@
 import React from 'react';
 import TestUtils from 'react-dom/test-utils';
 import { shallow } from 'enzyme';
-import { elementsTagTest, rootTagTest } from '../../utils/helpers/tags/tags-specs';
+import { rootTagTest } from '../../utils/helpers/tags/tags-specs';
 import AppWrapper from './app-wrapper';
 
 describe('app wrapper', () => {
-  let instance, wrapper;
+  let instance;
 
   beforeEach(() => {
-    instance = TestUtils.renderIntoDocument( <AppWrapper className='foobar'>foo</AppWrapper>);
-    wrapper = shallow(<AppWrapper className='foobar' data-element='app-wrapper' data-role='contacts'>foo</AppWrapper>);
+    instance = TestUtils.renderIntoDocument(<AppWrapper className='foobar'>foo</AppWrapper>);
   });
 
   it('renders the children', () => {
@@ -20,6 +19,12 @@ describe('app wrapper', () => {
   it('renders with correct classes', () => {
     let div = TestUtils.findRenderedDOMComponentWithTag(instance, 'div');
     expect(div.className).toEqual('carbon-app-wrapper foobar');
+  });
+
+  it('renders with additional html attributes', () => {
+    instance = TestUtils.renderIntoDocument(<AppWrapper style={{ color: 'red' }}>foo</AppWrapper>);
+    let div = TestUtils.findRenderedDOMComponentWithTag(instance, 'div');
+    expect(div.style.color).toEqual('red');
   });
 
   describe("tags", () => {