+++ src/components/button-toggle/__spec__.js
@@ -5,13 +5,10 @@ import { elementsTagTest, rootTagTest } from '../../utils/helpers/tags/tags-spec
 import ButtonToggle from './button-toggle';
 
 describe('ButtonToggle', () => {
-  let instance, wrapper;
+  let instance;
 
   beforeEach(() => {
     instance = TestUtils.renderIntoDocument(<ButtonToggle><span>Plus</span></ButtonToggle>);
-    wrapper = shallow(
-      <ButtonToggle data-role='contacts' data-element='button-toggle'><span>Minus</span></ButtonToggle>
-    );
   });
 
   describe('mainClasses', () => {
@@ -22,7 +19,7 @@ describe('ButtonToggle', () => {
 
   describe('input classes', () => {
     it('returns the classes for the input', () => {
-      expect(instance.inputClasses).toEqual('carbon-button-toggle__input hidden common-input__input');
+      expect(instance.inputClasses).toEqual('carbon-button-toggle__input common-input__input');
     });
   });