+++ src/utils/decorators/input-validation/input-validation.js
@@ -134,7 +134,8 @@ const InputValidation = ComposedComponent => class Component extends ComposedCom
 
   static contextTypes = assign({}, ComposedComponent.contextTypes, {
     form: PropTypes.object,
-    tab: PropTypes.object
+    tab: PropTypes.object,
+    modal: PropTypes.object
   });
 
   static propTypes = assign({}, ComposedComponent.propTypes, {
@@ -214,6 +215,7 @@ const InputValidation = ComposedComponent => class Component extends ComposedCom
    */
   componentWillMount() {
     // call the components super method if it exists
+    /* istanbul ignore else */
     if (super.componentWillMount) { super.componentWillMount(); }
 
     if (this.context.form && (this._validations() || this.props.warnings || this.props.info)) {
@@ -230,6 +232,7 @@ const InputValidation = ComposedComponent => class Component extends ComposedCom
    */
   componentWillUnmount() {
     // call the components super method if it exists
+    /* istanbul ignore else */
     if (super.componentWillUnmount) { super.componentWillUnmount(); }
 
     if (this._validations() || this.props.warnings || this.props.info) {
@@ -254,18 +257,30 @@ const InputValidation = ComposedComponent => class Component extends ComposedCom
 
       if (icon && message && message.offsetHeight) {
         let messagePositionLeft = (icon.offsetLeft + (icon.offsetWidth / 2));
-        const topOffset = icon.offsetTop - icon.offsetHeight;
+        const topOffset = icon.offsetTop - icon.offsetHeight,
+            messageOffsetWidth = message.offsetWidth;
 
         // set initial position for message
         message.style.left = `${messagePositionLeft}px`;
         message.style.top = `-${message.offsetHeight - topOffset}px`;
 
         // figure out if the message is positioned offscreen
-        const messageScreenPosition = message.getBoundingClientRect().left + message.offsetWidth;
+        const messageScreenPosition = message.getBoundingClientRect().left + messageOffsetWidth;
+
+        let shouldFlip = false;
 
         // change the position if it is offscreen
-        if (messageScreenPosition > this._window.innerWidth) {
-          messagePositionLeft -= message.offsetWidth;
+        if (this.context.modal && this.context.modal.getDialog()) {
+          // if in a modal check its position relative to that
+          const dialog = this.context.modal.getDialog();
+          shouldFlip = (message.offsetLeft + this._target.offsetLeft + messageOffsetWidth) > dialog.offsetWidth;
+        } else {
+          // otherwise check relative to the window
+          shouldFlip = messageScreenPosition > this._window.innerWidth;
+        }
+
+        if (shouldFlip) {
+          messagePositionLeft -= messageOffsetWidth;
           message.style.left = `${messagePositionLeft}px`;
           message.className += ' common-input__message--flipped';
           this.flipped = true;