+++ src/utils/logger/logger.js
@@ -1,10 +1,32 @@
 // Globally enable the logger
 let enabled = process.env.NODE_ENV !== 'production';
 
+const delays = {};
+const groupedMessages = {};
+
 // Log the message
-const log = (message, type) => {
+const log = (message, type, opts = {}) => {
   if (enabled) {
-    console[type](message); // eslint-disable-line no-console
+    if (opts.group) {
+      // if a group is defined, collect all occurrences to output together
+      groupedMessages[opts.group] = groupedMessages[opts.group] || [];
+      groupedMessages[opts.group].push(message);
+
+      const delay = delays[opts.group];
+
+      if (delay) {
+        clearTimeout(delay);
+      }
+
+      delays[opts.group] = setTimeout(() => {
+        console[type](groupedMessages[opts.group][0], { // eslint-disable-line no-console
+          all: groupedMessages[opts.group]
+        });
+      }, 500);
+    } else {
+      // output the message
+      console[type](message); // eslint-disable-line no-console
+    }
   }
 };
 
@@ -24,6 +46,9 @@ const log = (message, type) => {
  * warn - console.warn
  * deprecate - console.warn which prepends the message with [Deprecation]
  *
+ * You can also pass an option of 'group', this will group messages together that
+ * share the same group name and are triggered within 500ms of one another.
+ *
  * To Use:
  *
  * ```
@@ -38,24 +63,24 @@ const Logger = {
     enabled = newState;
   },
 
-  error: (message) => {
-    log(message, 'error');
+  error: (message, opts) => {
+    log(message, 'error', opts);
   },
 
-  info: (message) => {
-    log(message, 'info');
+  info: (message, opts) => {
+    log(message, 'info', opts);
   },
 
-  log: (message) => {
-    log(message, 'log');
+  log: (message, opts) => {
+    log(message, 'log', opts);
   },
 
-  warn: (message) => {
-    log(message, 'warn');
+  warn: (message, opts) => {
+    log(message, 'warn', opts);
   },
 
-  deprecate: (message) => {
-    log(`[Deprecation] ${message}`, 'warn');
+  deprecate: (message, opts) => {
+    log(`[Deprecation] ${message}`, 'warn', opts);
   }
 };